DUI: Reload SmartBar icons on boot complete
After a dirty flash or just a reboot in some cases, SmartBar
buttons that are intent actions with icons will show stock icon
instead of themed icon. This is because the bar is created and
loaded before CMTE is done processing the boot theme.

This is a temporary makeshift solution. In the future, query
theme service for processing themes. If CMTE is still processing,
register IThemeProcessingListener, and query again on callbacks.
When all themes are finished processing, unregister the listener
and poke navigation bar to update drawables.

Change-Id: I554aff4b5bf840ff83ea44de9dcc2572adf4b898